### 2018-09-22 #971 @whedon
[REVIEW]: piggyback: Managing Larger Data on a GitHub Repository
**Submitting author:** @cboettig (<a href="http://orcid.org/0000-0002-1642-628X">Carl Boettiger</a>)
**Repository:** <a href="https://github.com/ropensci/piggyback" target ="_blank">https://github.com/ropensci/piggyback</a>
**Version:** v0.0.6
