### 2018-09-22 #971 @annakrystalli
This submission has been accepted to rOpenSci. The review thread can be found at <https://github.com/ropensci/onboarding/issues/220>
