### 2018-09-23 #971 @arfon
Hey @cboettig - thanks for submitting this package. Do you think you could add a little more context to your paper including any examples of projects using this package?
