### 2018-09-23 #971 @cboettig
I’d be happy to! There’s a deal more in README and the two vignettes. In
the R world, we really think of vignettes as taking this role; i kinda
wonder if it would make more sense to replace paper.md with the intro
vignette? Or would that be too long for a paper.md? Or should these just be
linked instead?
On Sun, Sep 23, 2018 at 7:15 AM Arfon Smith <notifications@github.com>
wrote:

> Hey @cboettig <https://github.com/cboettig> - thanks for submitting this
> package. Do you think you could add a little more context to your paper
> including any examples of projects using this package?
>
