### 2018-09-23 #971 @cboettig
Right, it wasn’t clear to me if bullets 2 & 3 should involve actual code
calls. This briefer documentation is what we put in READMEs, (along with
quick install instructions).

Anyway, I’ll take a stab at this tonight . I still find this conceptually a
bit ambiguous; to me; anything that can be justified as being in paper.md
is probably meaningful enough to also already exist in part of the standard
package documentation. This advice id what we expect in readme and
vignettes in rOpenSci, because it is important like you say. I am a bit
unclear the extent to which paper.md should duplicate such information.
On Sun, Sep 23, 2018 at 1:20 PM Arfon Smith <notifications@github.com>
wrote:

> I think the vignettes are a little too much like technical documentation
> for the paper. The submission guidelines
> <https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain>
> list these as what your paper should contain:
>
>
>    - A list of the authors of the software and their affiliations
>    - A summary describing the high-level functionality and purpose of the
>    software for a diverse, non-specialist audience
>    - A clear statement of need that illustrates the purpose of the
>    software
>    - A list of key references including a link to the software archive
>    - Mentions (if applicable) of any ongoing research projects using the
>    software or recent scholarly publications enabled by it
>
> ... so a little bit more about high-level functionality, some use cases,
> and projects/research making use of this software already. Does that sound
> OK?
>
