### 2018-09-24 #971 @cboettig
@arfon Hey Arfon, I just took a stab at updating the `paper.md` along the lines you suggested (and tagged you in the commit).  Let me know if that's better?  Thanks for your patience and help on this, really appreciate the effort to get the right level of detail/content/focus in my `paper.md`'s, both for this time and future reference.
