### 2018-09-24 #971 @arfon
Thanks for the updates @cboettig - this is looking good!

Your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00971 :zap: :rocket: :boom:
