### 2018-09-25 #971 @arfon
> @arfon Thanks! Just a quick note: Looks like the "Software Archive" link at http://joss.theoj.org/papers/10.21105/joss.00971 is pointing back to the JOSS DOI, not the Zenodo DOI. Can we fix that?

That should be fixed now.
